pyproject.toml
[project]
name = "blog"
version = "0.2.1"
description = "the.dusktreader blog"
authors = [
    {name = "Tucker Beck", email ="tucker.beck@gmail.com"},
]
readme = "README.md"
license-files = [
    "LICENSE.md",
    "docs/LICENSE.md",
]
requires-python = "~=3.13"
dependencies = [
    "inflection>=0.5.1",
    "loguru>=0.7.3",
    "mkdocs-material>=9.6.9",
    "mkdocs-rss-plugin>=1.17.1",
    "mkdocstrings[python]>=0.29.0",
    "py-buzz>=7.0.0",
    "pydantic>=2.11.2",
    "pyyaml>=6.0.2",
    "snick>=1.4.1",
    "typer>=0.15.2",
]

[project.urls]
homepage = "https://blog.dusktreader.dev"
source = "https://github.com/dusktreader/blog"
changelog = "https://github.com/dusktreader/blog/blob/main/CHANGELOG.md"

[dependency-groups]
dev = [
    "ipython>=9.0.2",
    "pyclean>=3.1.0",
]

[project.scripts]
blog = "blog.main:cli"

[tool.uv]
package = true

[tool.pytest.ini_options]
addopts = [
    "--cov=src/blog",
    "--cov-report=term-missing",
    "--cov-fail-under=90",
    "--cov-report=xml:.coverage.xml",
    "--junitxml=.junit.xml",
    "--override-ini=junit_family=legacy",
]

[tool.ruff]
line-length = 120
src = ["src/blog", "tests"]

[tool.mypy]
pretty = true
files = ["src/blog", "tests"]

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true

[tool.basedpyright]
reportAny = false
reportExplicitAny = false
reportUnusedCallResult = false

[tool.hatch.build.targets.wheel]
packages = ["src/blog"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
